fix(offload): preserve namespaced model ids
